Split context and restructure
File context.rs is changed to wasm_context.rs.
execution in wasm_context are split into multiple files, each for 1 opcode.
Functions with opcode numbers are replaced with opcode names.
Create Struct ProofContext.
