+++ src/proof_context/trace/mod.rs
@@ -4,3 +4,4 @@ pub mod section_type;
 pub mod storage_type;
 pub mod storage_read_record;
 pub mod storage_write_record;
+pub mod trace_collector;
