+++ src/proof_context/trace/section_type.rs
