+++ src/proof_context/trace/state_trace_tuple.rs
@@ -1,11 +1,10 @@
-use crate::trace::storage_read_record::{
-    StorageReadRecord
-};
-use crate::trace::storage_write_record::StorageWriteRecord;
+use crate::proof_context::trace::storage_read_record::StorageReadRecord;
+use crate::proof_context::trace::storage_write_record::StorageWriteRecord;
 
 pub const MAX_NUM_READ_LOCATIONS: usize = 9;
 pub const MAX_NUM_WRITE_LOCATIONS: usize = 2;
 
+#[derive(Clone)]
 pub struct StateTraceTuple {
     pc_before_executing: u64,
     iaddr_before_executing: u64,
