+++ src/proof_context/trace/storage_type.rs
@@ -1,4 +1,4 @@
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub enum StorageType {
     Memory = 0,
     Stack = 1,
