+++ src/proof_context/trace/trace_collector/mod.rs
@@ -1,0 +1,3 @@
+pub mod opcode_end;
+mod opcode_local_get;
+mod opcode_i64_add;
