+++ src/proof_context/trace/trace_collector/opcode_i64_add.rs
@@ -1,0 +1,8 @@
+use crate::proof_context::proof_context::ProofContext;
+use crate::proof_context::trace::state_trace_manager::StateTraceManager;
+
+impl ProofContext {
+    pub fn collect_trace_opcode_i64_add(&mut self, pc: u64, iaddr: u64) {
+        todo!();
+    }
+}
