--- src/trace_collector/mod.rs
@@ -1,3 +1,0 @@
-pub mod opcode_0x0b;
-mod opcode_0x20;
-mod opcode_0x7c;
