--- src/trace_collector/opcode_0x7c.rs
@@ -1,7 +1,0 @@
-use crate::trace::state_trace_manager::StateTraceManager;
-
-impl StateTraceManager {
-    pub fn collect_0x7c(&mut self, time_stamp: &mut u64, pc: u64, iaddr: u64) {
-        todo!();
-    }
-}
