+++ src/util/mod.rs
@@ -1,2 +1,1 @@
-pub mod util;
 pub mod constant_setting;
